### 2026-02-15 #10011 @editorialbot
[PRE REVIEW]: Eegle: An open-source Julia integrative package for EEG data analysis and machine learning
**Submitting author:** @Marco-Congedo (<a href="http://orcid.org/0000-0003-2196-0409">Marco Congedo</a>)
**Repository:** https://github.com/Marco-Congedo/Eegle.jl
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.1.9
