### 2026-02-15 #10011 @Marco-Congedo
Suggested reviewers:
[Benedikt Ehinger](https://reviewers.joss.theoj.org/reviewers/2525)
[Isaac Kinley](https://reviewers.joss.theoj.org/reviewers/2093)
