+++ 0.2.3/siteinfo.js
@@ -1,2 +1,2 @@
-var DOCUMENTER_CURRENT_VERSION = "0.2.3";
+var DOCUMENTER_CURRENT_VERSION = "v0.2.3";
 var DOCUMENTER_IS_DEV_VERSION = false;
