### 2026-02-16 8397e6f
CompatHelper: add new compat entry for Statistics at version 1, (keep existing compat)
