### 2026-02-16 2870ae3
CompatHelper: add new compat entry for Downloads at version 1, (keep existing compat)
