+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Eegle"
 uuid = "1260424e-cd79-4b74-a0ea-47c69c296663"
-authors = ["Marco Congedo, Fahim Doumi"]
 version = "0.2.3"
+authors = ["Marco Congedo, Fahim Doumi"]
 
 [deps]
 ArtifactUtils = "8b73e784-e7d8-4ea5-973d-377fed4e3bce"
@@ -43,6 +43,7 @@ DSP = "0.8"
 DataFrames = "1.7"
 Diagonalizations = "0.2"
 Distributions = "0.25"
+Downloads = "1"
 EzXML = "1.2"
 FourierAnalysis = "v1.2"
 HDF5 = "0.17"
