### 2026-02-16 465153b
CompatHelper: add new compat entry for Artifacts at version 1, (keep existing compat)
