+++ versions.js
@@ -4,5 +4,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "0.2.3";
+var DOCUMENTER_NEWEST = "0.2.4";
 var DOCUMENTER_STABLE = "stable";
