+++ v0.2.6/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"documents/Benchmark/#BCI-FII-Corpus-Benchmark","page":"Benchmark","title":"BCI FII Corpus Benchmark","text":"This document reports the results of a benchmark run on the FII BCI Corpus using Eegle.\n\nOnly subjects/sessions that passed all exclusion criteria described in the @discarded.md file (see it for MI and P300) were included in this benchmark.\n\nIndividual subject/session accuracies are included in the .yml metadata files and are also available in CSV format in this GitHub repository. The repository also hosts the scripts used to make this benchmark and more information. \n\nThis document reports the mean(±sd) balanced accuracies for each database according to the analyzed tasks:     - MI: \"righthand\" vs. \"feet\" and \"lefthand\" vs. \"right_hand\"      - P300: target vs nontarget.","category":"section"},{"location":"documents/Benchmark/#Available-Classifiers","page":"Benchmark","title":"Available Classifiers","text":"The benchmark has employed three Riemannian classifiers, all adopting the affine-invariant (Fisher-Rao) metric:\n\nNative classifiers (acting on the manifold); the metric is used for computing distances and barycenters:\n\nMDM (Minimum Distance to Mean).\n\nTangent space classifiers; the metric is used to compute a barycenter for projecting the data onto the tangent space:\n\nENLR (Elastic Net Logistic Regression) \nLinear SVM (Linear Support Vector Machine).\n\n","category":"section"},{"location":"documents/Benchmark/#Processing-Pipelines","page":"Benchmark","title":"Processing Pipelines","text":"For both MI and P300 the default pipeline implementd by the crval function has bene used with the following two customizations:\n\nStratified K-Fold (K=10) instead of the default 8-fold\nSeed set to 109 (instead of default 1234).\n\n","category":"section"},{"location":"documents/Benchmark/#Motor-Imagery-(MI)-Benchmark-Results","page":"Benchmark","title":"Motor Imagery (MI) Benchmark Results","text":"","category":"section"},{"location":"documents/Benchmark/#Task:-right_hand-vs.-feet","page":"Benchmark","title":"Task: right_hand vs. feet","text":"details: show me\nDatabase MDM (%) ENLR (%) SVM (%)\nAlexMI-None 73.75 ± 18.03 80.62 ± 14.0 75.31 ± 13.39\nBNCI2014001-None 85.13 ± 12.2 87.7 ± 9.65 86.33 ± 11.05\nBNCI2014002-Test 80.77 ± 15.81 81.54 ± 16.88 81.92 ± 14.58\nBNCI2014002-Train 77.07 ± 11.4 77.54 ± 11.38 74.07 ± 13.87\nBNCI2015001-None 81.62 ± 13.08 84.53 ± 10.23 78.94 ± 12.42\nSchirrmeister2017-None 86.13 ± 12.67 96.48 ± 4.52 97.07 ± 3.62\nWeibo2014-None 61.94 ± 9.56 86.5 ± 8.9 81.72 ± 8.32\nZhou2016-None 87.33 ± 6.76 92.89 ± 3.66 87.45 ± 8.41","category":"section"},{"location":"documents/Benchmark/#Task:-left*hand-vs.-right*hand","page":"Benchmark","title":"Task: lefthand vs. righthand","text":"details: show me\nDatabase MDM (%) ENLR (%) SVM (%)\nBNCI2014001-None 77.75 ± 13.99 78.26 ± 15.33 78.85 ± 14.27\nBNCI2014004-Test 78.9 ± 12.41 80.9
[…]
