+++ .github/workflows/CI.yml
@@ -24,7 +24,7 @@ jobs:
         with:
           version: '1.11'
 
-      - uses: julia-actions/cache@v2
+      - uses: julia-actions/cache@v3
 
       - name: Install dependencies
         run: julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
