### 2026-03-30 #10011 @mooniean
Hi 👋 @kinleyid @behinger - would you be available and interested in reviewing this submission for the Journal of Open Source Software?

> Eegle: An open-source Julia integrative package for EEG data analysis and machine learning [#10011](https://github.com/openjournals/joss-reviews/issues/10011)

Thanks in advance!
