### 2026-04-09 #10388 @mooniean
Hi @Marco-Congedo,  @kinleyid, @AdamWysokinski 👋 - this is the review thread for the submission. All of our communications will happen here from now on.

Please check the post at the top of the issue for instructions on how to generate your own review checklist. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
