### 2026-05-29 948ff45
CompatHelper: bump compat for HTTP to 2, (keep existing compat)
