### 2026-06-01 deb415a
Update Julia section in paper.md for clarity
Clarified the description of active Julia projects for EEG data analysis. Answered a comment of AdamWysokinski
