+++ paper/paper.md
@@ -71,7 +71,7 @@ Older languages such as Python and Matlab have their established software ecosys
 
 Julia is a young open-source and cross-platform language specifically conceived for scientific computing [@julia:2017]. It is rapidly gaining momentum in the scientific community thanks to its conceptual affinity with mathematics and compatibility with the best available computing protocols. Although it is a high-level language, like Python and Matlab, it is (just-in-time) compiled, thus efficient. Moreover, Julia's syntax is elegant and permissive, allowing the programmer to adopt his/her preferred writing style. That is to say, the same routine in Julia can be written using a syntax closely resembling C, Python, or Matlab, to name a few. This accelerates the learning curve for people knowing other programming languages.
 
-For these reasons, the use of Julia may greatly benefit the field of EEG data analysis and machine learning. However, there are only a few active projects at the moment. The most active appear `NeuroAnalyzer.jl` [@Wysokinski:2024] and `unfold.jl` [@ehinger:2026]. The first focuses on sleep data and the second on event-related potentials. `Eegle.jl` is a generic package for EEG, thus its scope intersects only weakly with these specific packages.
+For these reasons, the use of Julia may greatly benefit the field of EEG data analysis and machine learning. However, there are only a few active projects at the moment. The most active appear `NeuroAnalyzer.jl` [@Wysokinski:2024] and `unfold.jl` [@ehinger:2026]. The first is a general package for importing/processing/analyzing and visualization of EEG and other neurophysiological data. The second focuses on event-related potentials. `Eegle.jl` is a generic package for EEG, thus its scope intersects somehow only with `NeuroAnalyzer.jl`, from which, however it differs by design.
 
 # Software design
 
