### 2026-06-02 19dd099
Refine summary of EEG and related tools
Updated the wording for clarity and precision in the summary section.
