+++ paper/paper.md
@@ -27,9 +27,9 @@ bibliography: paper.bib
 
 # Summary
 
-Existing since the 1920s, Electroencephalography (EEG) is the first non-invasive neuroimaging modality developed by mankind. Despite many more sophisticated modalities having been developed since, to date EEG is still by far the most widely used. This is due to a number of distinct advantages over other modalities, such as the high temporal resolution, the low price and encombrement of equipment, the silent operation and the total non-invasiveness.
+Existing since the 1920s, Electroencephalography (EEG) is the first non-invasive neuroimaging modality developed by mankind. Despite many more sophisticated modalities having been developed since, to date EEG is still by far the most widely used. This is due to a number of distinct advantages over other modalities, such as the high temporal resolution, the low price and portability of equipment, the silent operation and the total non-invasiveness.
 
-The recent explosion of research on EEG-based Brain-Computer Interfaces (BCIs) has fostered the need for efficient tools for EEG analysis and machine learning. While such tools exist for older languages such as Python and Matlab, they are not available for the more recent Julia language, which has been specifically created with these needs in mind. *Eegle.jl* leverages the rich and efficient scientific Julia ecosystem and integrates it to offer a simple and unified framework for both EEG data analysis and machine learning. In order to promote inter-operability of Julia and Python, we also release *pyLittleEegle*, a Python clone of the BCI-related capabilities found in *Eegle.jl*. Both packages work seamlessly with the *FII BCI Corpus*, the first large curated and annotated databases for the motor imagery and P300 BCI modalities.
+The recent explosion of research on EEG-based Brain-Computer Interfaces (BCIs) has fostered the need for efficient tools for EEG analysis and machine learning. While such tools exist for older languages such as Python and Matlab, they are not available for the more recent Julia language, which has been specifically created with these needs in mind. *Eegle.jl* leverages the rich and efficient scientific Julia ecosystem and integrates it to offer a simple and unified framework for both EEG data analysis and machine learning. In order to promote inter-operability of Julia and Python, we also release *pyLittleEegle*, a Python clone of the BCI-related capabilities found in *Eegle.jl*. Both packages work seamlessly with the *FII BCI Corpus*, the first large curated and annotated corpus of databases for the motor imagery and P300 BCI modalities.
 
 # Statement of need
 
