+++ v0.2.8/assets/Eegle_Braille.jl
@@ -1,0 +1,102 @@
+# https://505e06b2.github.io/Image-to-Braille/
+
+
+# Welcome Message
+print("                      🦅")
+println("\x1b[0m","
+⣴⣿⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛","\x1b[95m","⠛⠛⠛⠛⠛⠛","\x1b[0m","⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⠛⣿⣦","\x1b[35m", "
+⣿⣿","\x1b[38;5;231m","  𝑾  𝒆 𝒍 𝒄 𝒐 𝒎  𝒆   𝒕𝒐   𝑬 𝒆 𝒈 𝒍 𝒆 . 𝒋 𝒍","\x1b[35m","  ⣿⣿","\x1b[0m","
+⠻⣿⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤","\x1b[95m","⣤⣤⣤⣤⣤⣤","\x1b[0m","⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣤⣿⠟","\x1b[90m", "
+  https://github.com/Marco-Congedo/Eegle.jl
+","\x1b[0m")
+
+
+
+
+println("\x1b[38;5;6m","
+       ⣴⣿⣿⣿⣿⣿⣿⣿⣿⣿⠿⠟⠉","\x1b[38;5;172m","⠄⠊   ⠌  ⠜   ⠈⠄  ⠈⠄ ","\x1b[38;5;6m","⠸⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣦
[…]
