### 2026-06-25 133f513
CompatHelper: bump compat for Bonito to 5, (keep existing compat)
