### 2026-07-08 5db91a4
fix: Add CODECOV_TOKEN environment variable to Codecov upload step
