+++ .github/workflows/CI.yml
@@ -41,6 +41,8 @@ jobs:
           Pkg.test(coverage=true)'
 
       - name: Upload coverage to Codecov
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         run: julia -e '
           using Pkg;
           Pkg.add("Coverage");
