### 2026-07-09 71b595d
CompatHelper: bump compat for HTTP to 2, (keep existing compat)
