### 2026-07-09 7b41f1b
CompatHelper: bump compat for Bonito to 5, (keep existing compat)
