+++ .github/workflows/CI.yml
@@ -65,7 +65,7 @@ jobs:
         with:
           version: '1.11'
 
-      - uses: julia-actions/cache@v2
+      - uses: julia-actions/cache@v3
 
       - name: Install documentation dependencies
         run: julia --project=docs -e '
