+++ test/runtests.jl
@@ -2,6 +2,7 @@ using Eegle, Statistics, LinearAlgebra, Test
 
 tol = 1e-12
 
+include("test_FileSystem.jl")
 include("test_Preprocessing.jl")
 include("test_Processing.jl")
 include("test_ERPs.jl")
