+++ codecov.yml
@@ -1,13 +1,0 @@
-coverage:
-  status:
-    project:
-      default:
-        target: auto
-
-ignore:
-  - "docs/**"
-  - "test/**"
-  - "src/extras/**"
-  - "src/GUIs/**"
-  - "src/tools/Tyler.jl"
-  - "src/Database.jl"
