+++ .github/workflows/CI.yml
@@ -54,7 +54,8 @@ jobs:
           using Coverage
           LCOV.writefile("lcov.info", process_folder(pwd()))
           '
-          sed -i "s#\\\\#/#g" lcov.info
+          sed -i.bak 's#\\#/#g' lcov.info
+          rm lcov.info.bak
 
       - name: Upload coverage to Codecov
         if: matrix.os == 'ubuntu-latest' && matrix.version == '1.12'
