+++ test/test_Preprocessing.jl
@@ -11,6 +11,22 @@ println("\x1b[95m", "\nTesting module Eegle.Preprocessing.jl...", "\x1b[0m")
         stX = standardize(X; robust=true, prop=0.1) # execute only
 end;
 
+@testset "standardize" begin
+    X = randn(128, 19)
+
+    stX = standardize(X)
+
+    @test abs(mean(stX)) < tol
+    @test abs(var(stX) - 1) < tol
+
+    stXr = standardize(X; robust=true, prop=0.1)
+
+    @test size(stXr) == size(X)
+
+    # robust standardization should still roughly center the data
+    @test abs(mean(stXr)) < 0.2
+end
+
 ## resample (Downsampling tested and the other resampling cases executed only)
 @testset "resample                       " begin      
     sr = 128
@@ -24,30 +40,124 @@ end;
 
     Y=resample(X1, sr, 2) # upsample by a factor 2, i.e., double the sampling rate
     Y=resample(X1, sr, 100/sr) # downsample to 100 samples per second
+
+    @test X1 === resample(X1, sr, 1)    
+
+    stim = zeros(Int, size(X1,1))
+    stim[100] = 1
+    stim[500] = 2
+
+    Y, stim2 = resample(X1, sr, 1//2; stim=stim)
+
+    @test size(Y,1) == length(stim2)
+    @test sort(unique(stim2)) == [0,1,2]    
+
+    Y, stim2 = resample(X1, sr, 1; stim=stim)
+
+    @test Y === X1
+    @test stim2 === stim    
 end;
 
 ## removeChannels (one case is tested and the others are executed only)
 @testset "removeChannels                 " begin  
         X2 = randn(128, 7)
         sensors=["F7", "F8", "C3", "Cz", "C4", "P7", "P8"]
+        
         # remove second channel
         X2_, sensors_, ne = removeChannels(X2, 2, sensors)
         @test sensors_ == ["F7", "C3", "Cz", "C4", "P7", "P8"]
         @test norm(X2_ - hcat(X2[:, 1], X2[:, 3:end])) ≈ 0
+        
         # remove the first five channels
         X2_, sensors_, ne = removeChannels(X2, collect(1:5), sensors)
+        @test ne == 2
+        @test sensors_ == ["P7","P8"]
+        @test size(X2_) == (128,2)        
+        
         # remove the channel labeled as "Cz" in `sensors`
         X2_, sensors_, ne = removeChannels(X2, findfirst(x->x=="Cz", sensors), sensors)
+        @test ne == 6
+        @test sensors_ == ["F7", "F8", "C3", "C4", "P7", "P8"]
+        @test size(X2_) == (128,6)  
+        
         # remove the channels labeled as "C3", "Cz", and "C4" in `sensors`
         X2_, sensors_, ne = removeChannels(X2, findall(x->x∈("Cz", "C3", "C4"), sensors), sensors)
+        @test ne == 4
+        @test sensors_ == ["F7", "F8", "P7", "P8"]
+        @test size(X2_) == (128,4)          
+        
         # keep only channels labeled as "C3", "Cz", and "C4" in `sensors`
         X2_, sensors_, ne = removeChannels(X2, findall(x->x∉("Cz", "C3", "C4"), sensors), sensors)
+        @test ne == 3
+        @test sensors_ == ["C3", "Cz", "C4"]
+        @test size(X2_) == (128,3)  
+
 end;
 
+
+@testset "removeSamples" begin
+
+    X = float.(reshape(1:50,10,5))
+
+    stim = [0,1,0,2,0,3,0,0,1,0]
+
+    #
+    # remove one sample
[…]
