### 2026-07-29 727f1e8
trialsperclass and new zenodo links
