+++ src/Database.jl
@@ -102,7 +102,7 @@ struct InfoDB
     paradigm            :: String                           # experimental paradigm (MI, P300, etc.)
     files               :: Vector{String}                   # database's files 
     nSessions           :: Vector{Int}                      # sessions per subject 
-    nTrials             :: Dict{String, Vector{Int}}        # trials per class per session (class_name => [trials_per_session])
+    nTrials             :: Dict{String, Vector{Int}}        # trials per class per session (class_name => [trialspersession])
     nSubjects           :: Int                              # total number of subjects
     nSensors            :: Int                              # number of electrodes
     sensors             :: Vector{String}                   # name of sensors (Fz, Cz, etc.)
@@ -211,7 +211,7 @@ function infoDB(dbDir)
     labels              = typeof(info["stim"]["labels"])[]
     offset              = typeof(info["stim"]["offset"])[]
     nClasses            = typeof(info["stim"]["nclasses"])[]
-    nTrials             = typeof(info["stim"]["trials_per_class"])[]
+    nTrials             = typeof(info["stim"]["trialsperclass"])[]
 
     timestamp           = typeof(info["id"]["timestamp"])[]
     run                 = typeof(info["id"]["run"])[]
@@ -249,7 +249,7 @@ function infoDB(dbDir)
         push!(labels, stim["labels"])
         push!(offset, stim["offset"])
         push!(nClasses, stim["nclasses"])
-        push!(nTrials, stim["trials_per_class"])
+        push!(nTrials, stim["trialsperclass"])
 
         id = info["id"]
         push!(timestamp, id["timestamp"])
@@ -391,7 +391,7 @@ Supports shortcuts for common paths and automatic nested search.
 **Shortcuts:**
 - `sr` → `acquisition.samplingrate`
 - `ref` → `acquisition.reference`
-- `tpc` → `stim.trials_per_class`
+- `tpc` → `stim.trialsperclass`
 - `perfLHRH` → `perf.left_hand-right_hand`
 - `perfRHF` → `perf.right_hand-feet`
 =#
@@ -400,7 +400,7 @@ function _getNestedValue(data::Dict, path::String)
     shortcuts = Dict(
         "sr"        => "acquisition.samplingrate",
         "ref"       => "acquisition.reference",
-        "tpc"       => "stim.trials_per_class",
+        "tpc"       => "stim.trialsperclass",
         "perfLHRH"  => "perf.left_hand-right_hand",
         "perfRHF"   => "perf.right_hand-feet"
     )
@@ -491,7 +491,7 @@ inclusion = (("perf.left_hand-right_hand.MDM", x -> x >= 0.7),)
 # Available Shortcuts:
 - `sr` → `acquisition.samplingrate`
 - `ref` → `acquisition.reference`
-- `tpc` → `stim.trials_per_class`
+- `tpc` → `stim.trialsperclass`
 - `perfLHRH` → `perf.left_hand-right_hand`
 - `perfRHF` → `perf.right_hand-feet`
 
@@ -535,14 +535,14 @@ inclusion = (("acquisition.sensors", x -> length(x) >= 16),)
 # Specific number of electrodes
 inclusion = (("acquisition.sensors", x -> length(x) == 64),)
 
-**Dictionary Filters (trials_per_class, labels, perf)**
+**Dictionary Filters (trialsperclass, labels, perf)**
 
 # Minimum trials across all classes > 30
-inclusion = (("trials_per_class", x -> minimum(values(x)) > 30),)
[…]
