+++ Project.toml
@@ -42,7 +42,7 @@ Bonito = "4.1.11, 5"
 CSV = "0.10"
 CovarianceEstimation = "0.2"
 Coverage = "1.7.0"
-DSP = "0.7, 0.8"
+DSP = "0.8.5"
 DataFrames = "1.7"
 Diagonalizations = "0.2"
 Distributions = "0.25"
