### 2026-07-31 1ae00c0
Add DOI to existing bibliography entries
Added DOI fields for two articles in the bibliography.
