+++ paper/paper.bib
@@ -98,7 +98,8 @@
   journal = {Computational Intelligence and Neuroscience},
   year    = {2011},
   volume  = {2011},
-  pages   = {156869}
+  pages   = {156869},
+  doi     = {10.1155/2011/156869}
 }
 
 @article{mne:2013,
@@ -182,7 +183,8 @@
   journal = {Annual Review of Biophysics and Bioengineering},
   year    = {1973},
   volume  = {2},
-  pages   = {157--180}
+  pages   = {157--180},
+  doi     = {10.1146/annurev.bb.02.060173.001105}
 }
 
 
