### 2026-07-31 b4ba263
Add ORCID for author in paper metadata
Added ORCID for author Fahim Doumi.
