+++ paper/paper.md
@@ -13,6 +13,7 @@ authors:
     corresponding: true # (This is how to denote the corresponding author)
     affiliation: 1 
   - name: Fahim Doumi
+    orcid: 0009-0000-0014-8468
     equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
     affiliation: "1, 2" # (Multiple affiliations must be quoted)
 affiliations:
