### 2026-08-18 #10388 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] Below please find correct citation for Julia paper, including full title and DOI:
```
@article{doi:10.1137/141000671,
author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
title = {Julia: A Fresh Approach to Numerical Computing},
journal = {SIAM Review},
pages = {65-98},
doi = {10.1137/141000671},
eprint = {https://doi.org/10.1137/141000671}
}
```
- [ ] Similarly, see [here](https://scikit-learn.org/stable/about.html#citing-scikit-learn) for sklearn citation
- [ ] Wolpaw book has DOI available [here](https://academic.oup.com/book/1700)
- [ ] Brainstorm paper DOI [here](https://onlinelibrary.wiley.com/doi/10.1155/2011/879716)
- [ ] beginning of statement of need: add comma after 1893
- [ ] second paragraph of statement of need: remove "the" before "information transfer"
- [ ] bulleted list: remove hyphen in "century-long"
- [ ] next bullet point: replace "silentness" with "silence"
- [ ] in Python table: add spaces before citations and check broken reference
- [ ] Matlab table: spaces before citations
- [ ] second paragraph under Julia: replace "active" with "actively developed" and "appear" with "are"
- [ ] last sentence of paragraph: remove "somehow" and add comma after "however"
- [ ] first paragraph of Software design: add "a" before "foundational"
- [ ] You're inconsistent in your use of quotes (single vs. double) – if you're using American English, all single quotes should be double instead. If you're using British English, double quotes should be single quotes. I didn't notice any spellings that disambiguate one convention vs. the other in the rest of the paper.
