### 2026-08-20 d6e2ba0
Update bibliography entries with DOI
Added DOI and URL fields to several articles in the bibliography.
