+++ paper/paper.bib
@@ -21,7 +21,8 @@
   journal = {Computational Intelligence and Neuroscience},
   year    = {2011},
   volume  = {2011},
-  pages   = {879716}
+  pages   = {879716},
+  doi     = {10.1155/2011/879716}
 }
 
 @article{Caton:1875,
@@ -147,13 +148,16 @@
 }
 
 @article{julia:2017,
-  author  = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
-  title   = {A Fresh Approach to Numerical Computing},
-  journal = {SIAM Review},
-  year    = {2017},
-  volume  = {59},
-  number  = {1},
-  pages   = {65--98}
+author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
+URL = {https://doi.org/10.1137/141000671},
+eprint = {https://doi.org/10.1137/141000671}
 }
 
 @misc{pyLittleEegle:2025,
@@ -174,7 +178,8 @@
   journal={Journal of Machine Learning Research},
   volume={12},
   pages={2825--2830},
-  year={2011}
+  year={2011},
+  doi = {10.48550/arXiv.1201.0490}
 }
 
 @article{Vidal1973,
@@ -203,5 +208,6 @@
   title     = {Brain-Computer Interfaces: Principles and Practice},
   booktitle = {Brain-Computer Interfaces: Principles and Practice},
   publisher = {Oxford University Press},
-  year      = {2012}
+  year      = {2012},
+  doi       = {10.1186/1475-925X-12-22}
 }
