### 2026-08-20 4f708bc
Fix punctuation and formatting in paper.md
Corrected punctuation and formatting issues in the text, ensuring consistency in the presentation of information regarding EEG and BCI.
