### 2026-08-20 3dfd926
Change entry type from software to article in paper.bib
