+++ paper/paper.bib
@@ -114,7 +114,7 @@
   doi     = {10.3389/fnins.2013.00267}
 }
 
-@software{moabb:2025,
+@article{moabb:2025,
   author       = {Aristimunha, Bruno and
                   Carrara, Igor and
                   Guetschel, Pierre and
