### 2026-08-20 b390708
Fix typographical errors in paper.md
