### 2026-08-20 248472b
Add GitHub Actions workflow to compile JOSS paper
