### 2026-08-20 0724b02
Fix citation format for MOABB package
