+++ dev/assets/documentercitations/citations-data.js
@@ -1,0 +1,2 @@
+// Generated by DocumenterCitations: bibliography entry data for citation hover popups.
+window.DocumenterCitationsHoverData = {"zhang2016automatic":{"page":"references/index.html","html":"Zhang, T. and Wiesel, A. (2016). <a href=\"https://sciences.ucf.edu/math/tengz/wp-content/uploads/sites/45/2016/08/automatic-diagonal-loading-3.pdf\"><em>Automatic Diagonal Loading for Tyler&#39;s Robust Covariance Estimator<\/em><\/a>. In: <em>IEEE Statistical Signal Processing Workshop (SSP)<\/em>; pp. 1–5."},"Congedo2017Review":{"page":"references/index.html","html":"Congedo, M.; Barachant, A. and Bhatia, R. (2017). <a href=\"https://hal.science/hal-01570120/document\"><em>Riemannian geometry for EEG-based brain-computer interfaces; a primer and a review<\/em><\/a>. <a href=\"https://doi.org/10.1080/2326263X.2017.1297192\">Brain-Computer Interfaces <strong>4<\/strong>, 155–174<\/a>."},"congedo2008bss":{"page":"references/index.html","html":"Congedo, M.; Gouy-Pailler, C. and Jutten, C. (2008). <a href=\"https://hal.science/hal-00343628\"><em>On the blind source separation of human electroencephalogram by approximate joint diagonalization of second order statistics<\/em><\/a>. <a href=\"https://doi.org/10.1016/j.clinph.2008.07.014\">Clinical Neurophysiology <strong>119<\/strong>, 2677–2686<\/a>."},"GouyPailler2010":{"page":"references/index.html","html":"Gouy-Pailler, C.; Congedo, M.; Brunner, C.; Jutten, C. and Pfurtscheller, G. (2010). <em>Nonstationary Brain Source Separation for Multiclass Motor Imagery<\/em>. IEEE Transactions on Biomedical Engineering <strong>57<\/strong>, 469–478."},"Kim2023GhostICs":{"page":"references/index.html","html":"Kim, H.; Luo, J.; Chu, S.; Cannard, C.; Hoffmann, S. and Miyakoshi, M. (2023). <a href=\"https://doi.org/10.3389/frsip.2023.1064138\"><em>ICA’s bug: How ghost ICs emerge from effective rank deficiency caused by EEG electrode interpolation and incorrect re-referencing<\/em><\/a>. <a href=\"https://doi.org/10.3389/frsip.2023.1064138\">Frontiers in Signal Processing <strong>3<\/strong><\/a>. Section: Biomedical Signal Processing."},"Wiskott2002SFA":{"page":"references/index.html","html":"Wiskott, L. and Sejnowski, T. J. (2002). <em>Slow Feature Analysis: Unsupervised Learning of Invariances<\/em>. <a href=\"https://doi.org/10.1162/089976602317318938\">Neural Computation <strong>14<\/strong>, 715–770<\/a>."},"tyler1987":{"page":"references/index.html","html":"Tyler, D. E. (1987). <a href=\"https://projecteuclid.org/download/pdf_1/euclid.aos/1176350263\"><em>A Distribution-Free M-Estimator of Multivariate Scatter<\/em><\/a>. <a href=\"https://doi.org/10.1214/aos/1176350263\">The Annals of Statistics <strong>15<\/strong>, 234–251<\/a>."},"Molgedey1994TDSEP":{"page":"references/index.html","html":"Molgedey, L. and Schuster, H. G. (1994). <em>Separation of a Mixture of Independent Signals Using Time Delayed Correlations<\/em>. <a href=\"https://doi.org/10.1103/PhysRevLe
[…]
