### 2026-08-24 f196565
Bump EndBug/add-and-commit from 9 to 11
Bumps [EndBug/add-and-commit](https://github.com/endbug/add-and-commit) from 9 to 11.
- [Release notes](https://github.com/endbug/add-and-commit/releases)
- [Commits](https://github.com/endbug/add-and-commit/compare/v9...v11)

---
updated-dependencies:
- dependency-name: EndBug/add-and-commit
  dependency-version: '11'
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
