### 2026-08-24 ec40aba
Fix capitalization and improve text consistency
Corrected capitalization of 'MATLAB' and made minor text adjustments for consistency.
