### 2026-08-24 3c67eca
Fix: Add sudo to apt-get commands to resolve permission denied error
