### 2026-08-25 #10388 @Marco-Congedo
> Congrats to [@Marco-Congedo](https://github.com/Marco-Congedo) and thanks to handling editor [@mooniean](https://github.com/mooniean) and reviewers [@kinleyid](https://github.com/kinleyid) and [@AdamWysokinski](https://github.com/AdamWysokinski). JOSS wouldn't exist without your volunteer efforts!

Thank you very much for your work and assistance @rkurchin and @mooniean for the editorial role. Also, many thanks to the reviewers (@AdamWysokinski  and @kinleyid) , in particular the latter for the many nice suggestions.
